+++ changelog.md
@@ -1,7 +1,7 @@
 ## Lockout v0.11.7
 - Fixed an issue where mobs attached to non-players would count towards LeashUniqueEntities progress
 - Tweaked Pick/Ban chat messages to use the getGoalName string rather than the Goal ID string
-- Adjusted naming of "Right Click Banner with a Map" for better clarification
+- Adjusted naming of "Right Click Banner with Map" for better clarification
 - Updated Goal Requirement:
 	- Use Brush on Suspicious Block
 - Added New Biome Requirement: